+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `#sub_account_bnb_burn_status`
 - `#sub_account_set_spot_bnb_burn`
 
+### Changed
+- `Deposit#sub_account_id` is an Integer and no longer a String. Consistent with sub_account_id everywhere else
+
 ## [4.0.0]
 ### Added
 - Add optional `network` parameter when getting deposit address of a specific network